+++ packages/api/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @osdk/api
 
+## 2.13.0
+
 ## 2.12.0
 
 ### Minor Changes