+++ packages/cli.common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @osdk/cli.common
 
+## 0.37.0
+
 ## 0.36.0
 
 ### Minor Changes