+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/client.test.ontology
 
+## 2.13.0
+
+### Patch Changes
+
+- @osdk/api@2.13.0
+
 ## 2.12.0
 
 ### Patch Changes