+++ packages/example-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @osdk/example-generator
 
+## 0.19.0
+
+### Patch Changes
+
+- @osdk/create-app@2.13.0
+- @osdk/create-widget@3.10.0
+
 ## 0.18.0
 
 ### Patch Changes